quell warnings (part 4)
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21131 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/WeatherRenderer.cxx
@@ -429,7 +429,7 @@ void WeatherRenderer::update(void)
 {
   // update the time
   BzTime now = BzTime::getCurrent();
-  float frameTime = now - lastRainTime;
+  float frameTime = (float)(now - lastRainTime);
   lastRainTime = now;
 
   std::vector<rain> dropsToAdd;
